Make enter/heart button submit and add shift + enter for newlines
